Update changelog and docs for 0.2.0 and clean up code
The markdown protocol docs have been updated to reflect the changes
made for protobuf compatibility, and all the protobuf files have been
marked unstable.

To cleans things up a little bit the massive test for the proto
validation code has had some utility code factored out and some
function move out of the test.
