+++ test/modules/builder/builders/xviz-metadata-builder.spec.js
@@ -229,6 +229,7 @@ test('XVIZMetadataBuilder#stylesheet', t => {
   };
 
   t.deepEqual(metadata, expected, 'XVIZMetadataBuilder build matches expected output');
+  schemaValidator.validate('session/metadata', metadata);
   t.end();
 });
