+++ test/modules/schema/index.js
@@ -3,4 +3,5 @@ import './data.spec.js';
 import './examples.spec';
 import './file-validation.spec.js';
 import './proto-validation.spec.js';
+import './proto-utils.spec.js';
 import './validator.spec';
